refactor(frontend): isolate game features
Move game-specific views, hooks, tests, and CalPoker types into their feature directories so future terminal UI work has clear ownership boundaries.
