+++ front-end/src/features/calPoker/components/CaliforniaPoker.tsx
@@ -7,7 +7,7 @@ import {
   MovingCardData,
   cardIdToRankSuit,
   handValueToDescription,
-} from '../../../types/californiaPoker';
+} from '../types';
 // Constants
 import {
   GAME_STATES,
@@ -18,11 +18,11 @@ import {
 // Utils
 import { formatHandDescription, makeDescription, formatCardsForLog, formatOrderedCardsForLog, orderUsedCardsForLog } from './utils';
 import { HandDisplay, MovingCard } from './components';
-import { SuitName } from '../../../types/californiaPoker/CardValueSuit';
+import { SuitName } from '../types/CardValueSuit';
 import {
   CalpokerDisplaySnapshotView,
   CalpokerOutcomeView,
-} from '../../../types/californiaPoker/CaliforniapokerProps';
+} from '../types/CaliforniapokerProps';
 import GameBottomBar from './components/GameBottomBar';
 import {
   calpokerSettlementVerb,
