+++ front-end/src/features/calPoker/components/components/CardContent.tsx
@@ -1,4 +1,4 @@
-import { CardContentProps } from '../../../../types/californiaPoker';
+import { CardContentProps } from '../../types';
 import { RANK_SYMBOLS } from '../constants/constants';
 
 function CardContent(content: CardContentProps) {
