+++ front-end/src/features/calPoker/components/utils/gameLog.ts
@@ -1,4 +1,4 @@
-import { cardIdToRankSuit } from '@/src/types/californiaPoker';
+import { cardIdToRankSuit } from '../../types';
 
 const LOG_RANKS: Record<number, string> = {
   14: 'A', 13: 'K', 12: 'Q', 11: 'J', 10: 'T',
