+++ front-end/src/features/calPoker/types/CaliforniapokerProps.ts
@@ -37,5 +37,5 @@ export interface CaliforniapokerProps {
   initialSnapshot?: CalpokerDisplaySnapshotView;
   myName?: string;
   opponentName?: string;
-  settlementOutcome?: import('../../lib/settlement').SettlementOutcome | null;
+  settlementOutcome?: import('../../../lib/settlement').SettlementOutcome | null;
 }
