+++ front-end/src/features/calPoker/types/CardContentProps.ts
